Reorganize gardens mock data
