+++ mockServer/data/gardens/newGardenData.ts
@@ -1,0 +1,8 @@
+export const newGardenData = {
+  _id: "61376d19b4d3c1b26125f029",
+  name: "test6",
+  description: "",
+  fireBaseUserId: "tzFZEJMjR2TtGIN5Tk4poM7KMhA2",
+  createdDate: "2021-09-07T13:46:01.569Z",
+  lastUpdate: "2021-09-07T13:46:01.569Z",
+};
